# alon_sentinel/docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Alon Sentinel API
  version: 1.0.0
  summary: Stable public HTTP contract for Alon Sentinel v1.
  description: |
    Alon Sentinel is an API-first monitoring service. This contract mirrors
    the Axum route surface implemented under `src/api/routes`.
servers:
  - url: http://127.0.0.1:3000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /v1/heartbeat/{heartbeat_token}:
    parameters: [ { $ref: "#/components/parameters/HeartbeatToken" } ]
    post:
      operationId: recordHeartbeatPing
      tags: [Monitoring]
      summary: Record a heartbeat ping
      security: []
      responses: { "204": { description: Heartbeat recorded }, "404": { $ref: "#/components/responses/NotFound" } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: opaque }
...
  parameters:
...
    HeartbeatToken: { name: heartbeat_token, in: path, required: true, schema: { type: string } }
...
  responses:
...
    NotFound: { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties: { error: { type: string } }
...